+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [Unreleased] - 2017-04-28
 ### Added
 - Added a CHANGELOG.md (Yay!!)
+- Minor refactoring of mergeTypes
 
 ## [0.0.10] - 2017-04-27
 ### Fixed